# protocol/patricia: avoid allocs in node.Hash